requirements.txt
streamlit
python-dotenv
pydeck
pandas
requests
wikidata
plotly
pydantic